feat(wallet-sdk): add cross chain complete subscribe method
