+++ packages/apps/wallet-sdk-example/src/hooks/transfers.ts
@@ -85,6 +85,7 @@ export const useTransfers = () => {
     /* -- End demo ---------------*/
 
     walletSdk.subscribeOnCrossChainComplete(
+      wallet.account?.name ?? '',
       incompleteTransfers,
       () => refetch(),
       {
